requirements.txt
django
djangorestframework
drf-spectacular
pytest
pytest-django